Refactor list books and create book
